// src/api/models/staffListItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { StaffListStoreItem } from './staffListStoreItem';

export interface StaffListItem {
  /** ID */
  id: string;
  /** 이름 */
  name: string;
  /**
   * 이메일
   * @nullable
   */
  email?: string | null;
  /** 전화번호 */
  phoneNumber: string;
  /** 생성일 */
  createdAt: string;
  /** 등록된 매장 목록 */
  stores: StaffListStoreItem[];
}